Consistently use var
This comment also applies to the generated parser itself, but I'm
illustrating with the JSON example runner because all the variations
are visible within a short section of code: (1) const/let, (2) var
and (3) the absence of any such keyword.

Mixing (1) and (2) is frowned upon; it makes the scope of variables
unpredictable. (1) is newer, hence often considered better. In this
case, however, I would suggest that you consistently go with (2), at
least in the generated parser module. There are two reasons for my
recommendation:

 1. var has roughly the same scope semantics as local variables in
    Python. Since you are transpiling from Python, this makes it
    easier to generate reliable code.
 2. var is older, hence portable to more different JS environments.

(3), while allowed, should be avoided at all cost, as it always
creates global variables (which is not the same as var at module
scope). The fact that this form is allowed at all should be
considered a historical artifact.