Docstrings
Using multiline strings as documentation is not a thing in JS; you
are expected to always use comments, which should precede the thing
they document. The style I've demonstrated in the commit could be
broadly referred to as a literate commenting style, which is for
example practiced in Underscore and related libraries. It can be read
as-is together with the code, or it can be used to present the
comments and the code side by side, as illustrated below. Needless to
say, this necessitates writing a separate manual by hand (second
link).

https://underscorejs.org/docs/underscore-esm.html
https://underscorejs.org

A possible alternative style, which is very popular, is the jsdoc
format:

/**
 * Short description
 * Possibly longer explanation with examples
 * @param name description
 * @returns description
 */

This is more geared to automatically generating reference
documentation that is separate from the code. In order for such
separate documentation to be useful, the comments need to be quite
extensive, which can render the code itself less readable. It should
also be noted that the generated reference documentation is generally
still not sufficient to completely replace a purpose-written manual.

https://github.com/lodash/lodash/blob/c84fe82760fb2d3e03a63379b297a1cc1a2fce12/lodash.js#L459