Indentation
Two-space indents are often considered a convention in JavaScript.
Even in the JS community, however, there is sizable subpopulation that
appreciates larger indents for better readability. Ultimately, it is
important not to do what is most popular but what you believe is best.
If you subscribe to PEP8's four-space indent recommendation, there is
no reason not to apply it in JS as well.

(The reason two-space indents became fashion in JS is also outdated.
When JS programmers were still writing module wrappers by hand and
there was no async/await syntax yet, indentation tended to go very
deep very quickly. This produced an incentive to sacrifice readability
in exchange for needing to wrap lines less often. There is no longer
a need to make this sacrifice.)