Use speaking names
When possible, avoid abbreviated names because they cause cognitive
strain. Highly idiomatic single-letter variables such as i/j/k in a
loop can be acceptable.

An additional remark on this snippet: it appears that the re_
parameter is just an alias of the Python re module emulation a few
lines up. Avoid this pattern if possible; using re directly will make
the code much easier to understand.