Generator functions
JavaScript's modern iterator support is of course convenient for
transpiling Python generators. Unfortunately, they do have a
performance cost. for-of loops also work on plain arrays, so when it
is possible to replace a generator function by a normal function that
returns a plain array, it is likely that this will make your parser
faster.

for-of loops on arrays, in turn, should be replaced by old-fashioned
for-;; loops for raw performance. The for-of construct involves
converting the array to iterator first through the Array[@@iterator]
method and then calling iterator#next on each loop iteration, inducing
considerable function call overhead.

Of course, these considerations only apply to hot loops.

An added bonus of reducing iterators and for-of constructs is that
compatibility transpilers like Babel will produce less boilerplate
code.